Ardilla voladora v2
-Arreglados parcialmente los problemas físicos con la ardilla voladora (ya no se lleva a Nico por el escenario)
-Implementado el campo de visión "circular"
-Cuando está enfrente de un obstáculo se da la vuelta, pero por alguna razón no funciona con el "Ground"

* Hay muchas innecesarias en el código que luego voy a quitar
* Estaré agregando comentarios para los próximos commit

+++ Game/FinalProject/Assets/Scripts/Entities/Player/PlayerManager.cs
@@ -255,7 +255,7 @@ public class PlayerManager : Entity
     }
 
     #region Self state methods
-    public void Captured(int nTaps, int damagePerSecond)
+    public void Captured(int nTaps, float damagePerSecond)
     {
         if(!tirementRunning)
         {
